requirements.txt
appnope==0.1.2
backcall==0.2.0
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
entrypoints==0.3
fonttools==4.28.5
ipykernel==6.7.0
ipython==7.31.1
jedi==0.18.1
jupyter-client==7.1.2
jupyter-core==4.9.1
kiwisolver==1.3.2
matplotlib==3.5.1
matplotlib-inline==0.1.3
nest-asyncio==1.5.4
numpy==1.21.5
opencv-python==4.5.5.62
packaging==21.3
pandas==1.3.5
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.0
prompt-toolkit==3.0.24
ptyprocess==0.7.0
Pygments==2.11.2
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
pyzmq==22.3.0
six==1.16.0
tornado==6.1
traitlets==5.1.1
wcwidth==0.2.5
